+++ lib/src/code_generators/swagger_converter_generator.dart
@@ -20,7 +20,7 @@ ${_generateModelsMapping(dartCode, options)}};''';
     final classes = getDefinitions(map);
 
     final classesFromResponses =
-        SwaggerModelsGenerator.getClassesFromResponses(dartCode);
+        SwaggerModelsGenerator.getClassesFromResponses(dartCode, options);
 
     classes.addAll(classesFromResponses);
 
@@ -74,7 +74,7 @@ ${_generateModelsMapping(dartCode, options)}};''';
           if (schema != null &&
               content!.entries.length == 1 &&
               !schema.containsKey('\$ref')) {
-            final validatedName = key.capitalize;
+            final validatedName = key.capitalize + options.modelPostfix;
             result.add('\t$validatedName: $validatedName.fromJsonFactory,');
           }
         }