+++ lib/src/models/generator_options.dart
@@ -22,6 +22,8 @@ class GeneratorOptions {
     this.useInheritance = true,
     this.includeIfNull,
     this.modelPostfix = '',
+    this.includePaths = const [],
+    this.excludePaths = const [],
   });
 
   /// Build options from a JSON map.
@@ -50,7 +52,7 @@ class GeneratorOptions {
   final bool usePathForRequestNames;
 
   @JsonKey(defaultValue: null)
-  final IncludeIfNull? includeIfNull;
+  final bool? includeIfNull;
 
   @JsonKey(defaultValue: '')
   final String inputFolder;
@@ -76,6 +78,12 @@ class GeneratorOptions {
   @JsonKey(defaultValue: <ResponseOverrideValueMap>[])
   final List<ResponseOverrideValueMap> responseOverrideValueMap;
 
+  @JsonKey(defaultValue: [])
+  final List<String> includePaths;
+
+  @JsonKey(defaultValue: [])
+  final List<String> excludePaths;
+
   /// Convert this options instance to JSON.
   Map<String, dynamic> toJson() => _$GeneratorOptionsToJson(this);
 }
@@ -96,22 +104,6 @@ class DefaultValueMap {
 
   /// Convert this default value map instance to JSON.
   Map<String, dynamic> toJson() => _$DefaultValueMapToJson(this);
-}
-
-@JsonSerializable(fieldRename: FieldRename.snake)
-class IncludeIfNull {
-  IncludeIfNull({required this.enabled, required this.value});
-
-  @JsonKey(defaultValue: false)
-  final bool enabled;
-
-  @JsonKey(defaultValue: false)
-  final bool value;
-
-  factory IncludeIfNull.fromJson(Map<String, dynamic> json) =>
-      _$IncludeIfNullFromJson(json);
-
-  Map<String, dynamic> toJson() => _$IncludeIfNullToJson(this);
 }
 
 @JsonSerializable(fieldRename: FieldRename.snake)